petishh/requirements.txt
asgiref==3.4.1
coverage==6.1.2
Django==3.2.9
django-cors-headers==3.10.0
django-cors-middleware==1.5.0
django-filter==21.1
django-versatileimagefield==2.2
djangorestframework==3.12.4
djangorestframework-simplejwt==5.0.0
importlib-metadata==4.8.2
Markdown==3.3.6
mysqlclient==2.1.0
Pillow==8.4.0
PyJWT==2.3.0
PyMySQL==1.0.2
python-magic==0.4.24
python-magic-bin==0.4.14
pytz==2021.3
slugify==0.0.1
sqlparse==0.4.2
zipp==3.6.0
